+++ frontend/pages/admin/data/garbage-collection/index.tsx
@@ -28,7 +28,7 @@ import { getBuildingsOfTour } from "@/lib/tour";
 import { withAuthorisation } from "@/components/withAuthorisation";
 import BuildingAutocomplete from "@/components/autocompleteComponents/buildingAutocomplete";
 import TourAutocomplete from "@/components/autocompleteComponents/tourAutocomplete";
-import BulkOperationModal from "@/components/garbage/BulkOperationModal";
+import BulkMoveGarbageModal from "@/components/garbage/BulkMoveGarbageModal";
 
 interface ParsedUrlQuery {}
 
@@ -63,7 +63,7 @@ function GarbageCollectionSchedule() {
     const [showBuildingListModal, setShowBuildingListModal] = useState<boolean>(false);
 
     // The bulk operation modal
-    const [showBulkOperationModal, setShowBulkOperationModal] = useState<boolean>(false);
+    const [showBulkMoveModal, setShowBulkMoveModal] = useState<boolean>(false);
 
     const [buildingList, setBuildingList] = useState<BuildingInterface[]>([]);
 
@@ -288,7 +288,7 @@ function GarbageCollectionSchedule() {
     }
 
     function closeBulkOperationModal() {
-        setShowBulkOperationModal(false);
+        setShowBulkMoveModal(false);
         getFromRange(currentRange);
     }
 
@@ -318,10 +318,10 @@ function GarbageCollectionSchedule() {
                 show={showDuplicateModal}
                 buildings={buildingList}
             />
-            <BulkOperationModal
+            <BulkMoveGarbageModal
                 buildings={buildingList}
                 closeModal={closeBulkOperationModal}
-                show={showBulkOperationModal}
+                show={showBulkMoveModal}
             />
             <GarbageEditModal
                 closeModal={closeEditModal}
@@ -350,7 +350,7 @@ function GarbageCollectionSchedule() {
                         </Button>
                     </div>
                     <div className="col">
-                        <Button variant="primary" className="btn-dark" onClick={() => setShowBulkOperationModal(true)}>
+                        <Button variant="primary" className="btn-dark" onClick={() => setShowBulkMoveModal(true)}>
                             Verplaats ophaling
                         </Button>
                     </div>
