--- frontend/pages/admin/data/schedules/edit.tsx
@@ -1,14 +1,0 @@
-import AdminHeader from "@/components/header/adminHeader";
-
-export default function AdminDataScheduleEdit() {
-    return (
-        <>
-            <>
-                <AdminHeader />
-                <p>
-                    https://www.figma.com/proto/9yLULhNn8b8SlsWlOnRSpm/SeLab2-mockup?node-id=115-205&scaling=contain&page-id=0%3A1&starting-point-node-id=118%3A1486
-                </p>
-            </>
-        </>
-    );
-}
