+++ frontend/pages/user/profile.tsx
@@ -9,8 +9,9 @@ import SyndicHeader from "@/components/header/syndicHeader";
 import { handleError } from "@/lib/error";
 import PasswordModal from "@/components/password/passwordModal";
 import ErrorMessageAlert from "@/components/errorMessageAlert";
+import {withAuthorisation} from "@/components/withAuthorisation";
 
-export default function UserProfile() {
+function UserProfile() {
     const { t } = useTranslation();
     const [user, setUser] = useState<User | null>(null);
     const [firstName, setFirstName] = useState<string>("");
@@ -228,3 +229,5 @@ export default function UserProfile() {
         </>
     );
 }
+
+export default withAuthorisation(UserProfile, ["Admin", "Superstudent", "Syndic", "Student", "Default"]);
